k
ok

+++ JAVA_SHOP/src/entities/Product.java
@@ -1,10 +1,10 @@
 package entities;
 
 public class Product {
-	private float price,tax;
+	private double price,tax;
 	private String title,category,description;
 	
-	public Product(float aprice,float atax,String atitle,String aCategory,String aDescription) {
+	public Product(double aprice,double atax,String atitle,String aCategory,String aDescription) {
 		this.price = aprice;
 		this.tax = atax;
 		this.title = atitle;
@@ -12,13 +12,13 @@ public class Product {
 		this.description = aDescription;
 	}
 	
-	public float getPrice() {
+	public double getPrice() {
 		return price = price + price * tax;
 	}
 	public void setPrice(float price) {
 		this.price = price;
 	}
-	public float getTax() {
+	public double getTax() {
 		return tax;
 	}
 	public void setTax(float tax) {
